threads: Solve 06_interior_mutability
This solution is good enough for this chapter/exercise, but it does
not *actually* do what its supposed to do. If the Rc reference count
goes down to 0, the shared memory on the heap is deallocated, meaning
new object creations (after dropping all previous objects) will reset
the drop counting to 0, since new memory will be allocated on the heap.